refactor(division-gateway): gateway 분할, service 추가
