+++ src/gateway/gateway.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuctionGateway } from './gateway';
-import { AuctionGatewayService } from './gateway.service';
+import { AuctionItemBidService } from './item/auctionItemBid.service';
 import { AuctionService } from '~/src/domain/auction/auction.service';
 import { AuctionManager } from '~/src/domain/auction/auction.manager';
 import { AuctionItemRepository } from '~/src/domain/auction/auction-item.repository';
@@ -10,6 +9,13 @@ import { AuctionItem } from '~/src/domain/auction/entities/auction-item.entity';
 import { User } from '~/src/domain/users/entities/user.entity';
 import { AuctionModule } from '~/src/domain/auction/auction.module';
 import { UsersModule } from '~/src/domain/users/users.module';
+import { AuctionChatGateway } from '~/src/gateway/chat/auctionChat.gateway';
+import { AuctionChatService } from '~/src/gateway/chat/auctionChat.service';
+import { AuctionItemBidGateway } from '~/src/gateway/item/auctionItemBid.gateway';
+import { AuctionJoinGateway } from '~/src/gateway/join/auctionJoin.gateway';
+import { AuctionJoinService } from '~/src/gateway/join/auctionJoin.service';
+import { AuctionNotificationService } from '~/src/gateway/notification/auctionNotification.service';
+import { AuctionNotificationGateway } from '~/src/gateway/notification/auctionNotification.gateway';
 
 @Module({
   imports: [
@@ -18,12 +24,18 @@ import { UsersModule } from '~/src/domain/users/users.module';
     UsersModule,
   ],
   providers: [
-    AuctionGateway,
-    AuctionGatewayService,
+    AuctionChatGateway,
+    AuctionChatService,
+    AuctionItemBidGateway,
+    AuctionItemBidService,
+    AuctionJoinGateway,
+    AuctionJoinService,
+    AuctionNotificationGateway,
+    AuctionNotificationService,
+    AuctionItemRepository,
     AuctionService,
     AuctionManager,
-    AuctionItemRepository,
   ],
-  exports: [AuctionGatewayService],
+  exports: [AuctionService],
 })
 export class GatewayModule {}
